Tech data access has been heavily optimised
Previously to access a tech data field you either had to look in the
cache for it or iterate the
whole table (around 1000 entries)

You always iterated the table when looking for a mapname's corresponding
tech id. <- This operation is done each time a class with TechMixin is
instantiated, like the ScriptActor class. Fun.