Make GUIManager predict the future.
Before executing a script, GUIManager checks the max time the script has
ever used to update.
It then checks if it *could* take too much time for a single frame, and
then decides based on that. It is not perfect, but it is good enough.

Furthermore, no longer does there exist an update interval.
This means that scripts will be executed as often as possible while
staying under the max update time.

The max update time has also been changed to 8 ms, since it is much more
likely to stay under this value, compared to before.